requirements.txt
asgiref==3.5.0
astor==0.8.1
Babel==2.9.1
backports.zoneinfo==0.2.1
bce-python-sdk==0.8.64
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
colorlog==6.6.0
cycler==0.11.0
decorator==5.1.1
dill==0.3.4
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
easydict==1.9
filelock==3.4.2
flake8==4.0.1
Flask==2.0.2
Flask-Babel==2.0.0
fonttools==4.29.1
future==0.18.2
gunicorn==20.1.0
h5py==3.6.0
identify==2.4.8
idna==3.3
imageio==2.15.0
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
multiprocess==0.70.12.2
networkx==2.6.3
nodeenv==1.6.0
numpy==1.21.3
onnx==1.9.0
opencv-python==4.5.4.58
packaging==21.3
paddle2onnx==0.9.0
paddlehub==2.2.0
paddlenlp==2.2.4
paddlepaddle==2.2.2
pandas==1.4.0
Pillow==9.0.1
platformdirs==2.4.1
pre-commit==2.17.0
protobuf==3.19.4
pycodestyle==2.8.0
pycryptodome==3.14.1
pyflakes==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
rarfile==4.0
requests==2.27.1
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.8.0
seqeval==1.2.2
shellcheck-py==0.8.0.4
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tifffile==2022.2.9
toml==0.10.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.13.1
visualdl==2.2.3
Werkzeug==2.0.2
